Make changes to fix depedency cycle error
