+++ js/project-list.js
@@ -1,0 +1,92 @@
+const projectList = [
+  {
+    id: 1,
+    title: 'tonic',
+    featuredImage: {
+      url: 'images/snapshoot-project-1.png',
+      alt: 'snapshoot-project-1',
+    },
+    description: {
+      less: 'A daily selection of privately personalized reads; no accounts or sign-ups required.',
+      more: 'Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry\'s standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essent.',
+    },
+    technologies: ['HTML', 'CSS', 'JavaScript'],
+    projectDetail: {
+      client: 'canopy',
+      role: 'back end dev',
+      year: 2015,
+    },
+    link: {
+      source: '',
+      live: '',
+    },
+  },
+  {
+    id: 2,
+    title: 'multi-post stories',
+    featuredImage: {
+      url: 'images/snapshoot-project-2.png',
+      alt: 'snapshoot-project-2',
+    },
+    description: {
+      less: 'A daily selection of privately personalized reads; no accounts or sign-ups required.',
+      more: 'Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry\'s standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essent.',
+    },
+    technologies: ['HTML', 'CSS', 'JavaScript'],
+    projectDetail: {
+      client: 'canopy',
+      role: 'back end dev',
+      year: 2015,
+    },
+    link: {
+      source: '',
+      live: '',
+    },
+  },
+  {
+    id: 3,
+    title: 'tonic',
+    featuredImage: {
+      url: 'images/snapshoot-project-3.png',
+      alt: 'snapshoot-project-3',
+    },
+    description: {
+      less: 'A daily selection of privately personalized reads; no accounts or sign-ups required.',
+      more: 'Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry\'s standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essent.',
+    },
+    technologies: ['HTML', 'CSS', 'JavaScript'],
+    projectDetail: {
+      client: 'canopy',
+      role: 'back end dev',
+      year: 2015,
+    },
+    link: {
+      source: '',
+      live: '',
+    },
+  },
+  {
+    id: 4,
+    title: 'multi-post stories',
+    featuredImage: {
+      url: 'images/snapshoot-project-4.png',
+      alt: 'snapshoot-project-3',
+    },
+    description: {
+      less: 'A daily selection of privately personalized reads; no accounts or sign-ups required.',
+      more: 'Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry\'s standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It has survived not only five centuries, but also the leap into electronic typesetting, remaining essent.',
+    },
+    technologies: ['HTML', 'CSS', 'JavaScript'],
+    projectDetail: {
+      client: 'canopy',
+      role: 'back end dev',
+      year: 2015,
+    },
+    link: {
+      source: '',
+      live: '',
+    },
+  },
+];
+
+export default projectList;
